Added ObjectContainerAcces class which optimizes build speed and solves future problems
Due to some problems width ObjectContainer in game branch, mostly concerning the use of template classes, it was decided to add a helper class for accessing member classes of world. This new class is purely static and #includes no other files while still having full access to ObjectContainer. As far as possible all #includes of ObjectContainer were replaced with #include ObjectContainerAccess. Note the use of #define, should maybe be standardized if used more in eng. Also the amount of needed friends of objectContainer was reduced.